Password protection for boards
This commit adds the ability to password protect boards. Users can now join a protected board by entering the password, and once joined, they do not need to provide the password again. If a board is password protected, users can view it in read-only mode, but cannot perform any actions that would edit it, and will get an alert if they try. The lock symbol now indicates if a board is password protected or not, and users can set or update a password by clicking on it.

Additionally, if a board is in read-only mode, clicking on the lock icon allows the user to enter the password and unlock the board to activate editing functionalities. Note that even after unlocking the board, users still need to enter the password in board overview to fully join it.